Implementation of the exception propagation by PIP
PIP propagates interrupts and faults, which are grouped by ARM under
the generic term Exceptions, in two different ways:

Faults, such as HardFault, are propagated to the parent partition of
the faulty partition. If the index corresponding to the fault number in
the VIDT of the parent partition contains a null value or a pointer to
an invalid context, the fault is propagated to the grandparent
partition, then to the great-grandparent partition and, finally, to the
root partition. If the root partition is unable to handle the fault, PIP
is in an unrecoverable state.

Interrupts, such as SysTick, are propagated to the root partition, which
can propagate interrupts to its children.