Change the physical MPU regions replacement policy
PIP provides a virtualization mechanism of the physical MPU, allowing
the creation of virtual MPU regions that do not respect the constraints
specific to ARMv7-M. If a MemManage exception is raised, PIP looks if
the faulted address is contained in one of the virtual MPU regions. If
this is the case, PIP searches for the largest physical MPU region
containing the faulted address. The previous physical MPU region
replacement policy was to replace the old physical MPU region with the
new one at the virtual MPU region index.

This policy does not work if:

  1. An instruction aligned on 2 bytes and coded on 4 bytes is located
     between 2 physical MPU regions;

  2. An instruction accesses an address that is in its virtual MPU
     region but not in its physical MPU region;

  3. An instruction accesses an address between 2 physical MPU regions;

  4. The 1st and 3rd combined.

The new physical MPU region replacement policy is to alternate between:

  1. The index of the virtual MPU region and index 6 for MemManage
     caused by an IACCVIOL;

  2. The index of the virtual MPU region and index 7 for MemManage
     caused by an DACCVIOL.

This should fix the problematic cases mentioned above at the cost of
losing two MPU regions.