Cut the RAM into 2 MPU regions
The RAM is now divided into 2 MPU regions:

The first one is a 32K physical region of RAM, which respects the
ARMv7-M constraints on the MPU. In this way, this MPU region will never
be unmapped by the MPU virtualization mechanism of Pip. It contains,
among other things, the stack and the VIDT of the root partition.

The second one is a virtual region containing the remaining available
RAM to the root partition.