Change the behavior when a call to Pip_in() or Pip_out() fails
Previously, Pip_in() and Pip_out() system calls returned a value to the
calling partition, indicating whether the read/write had failed. This
behavior does not match the philosophy of Pip. Indeed, it is up to the
parent partition to decide what to do when one of its children fails.

Now, Pip_in() and Pip_out() system calls trigger a MemManage fault on
failure, which is propagated to the parent partition.