Change the write permission for the MPU region 2
Region 2 of the MPU, which corresponds to the flash memory area
allocated to the root partition, was configured by Pip to prevent
partitions from rewriting their flash memory.

Pip now initialises MPU region 2 with read, write and execute
permissions.

This change is necessary if we want a partition to be able to load code
into its flash memory on the fly.