Fix a bug in the freeSlot() function
To erase a blockEntry_t structure, the freeSlot() function called the
writeBlockEntryFromBlockEntryAddr() function with default values. In
particular, it used the KERNELSTRUCTUREENTRIESNB value retrieved by the
getKernelStructureEntriesNb() function as the default value for the
index parameter. This value is not a valid index number because it is
not between [0;KERNELSTRUCTUREENTRIESNB[.

The problem is that some functions that follow the call to the
writeBlockEntryFromBlockEntryAddr() function call the
getKernelStructureStartAddr() function, which uses this index to
calculate the address of the kernel structure from the address of a
blockEntry_t. This function then returns the wrong kernel structure
address.

To correct this, we use the current index of the blockEntry_t as the
default value for the index parameter.